Dry Unique and Exists rules.
