+++ src/Illuminate/Validation/Rules/Unique.php
@@ -2,23 +2,9 @@
 
 namespace Illuminate\Validation\Rules;
 
-use Closure;
-
 class Unique
 {
-    /**
-     * The table to run the query against.
-     *
-     * @var string
-     */
-    protected $table;
-
-    /**
-     * The column to check for uniqueness on.
-     *
-     * @var string
-     */
-    protected $column;
+    use DatabaseRule;
 
     /**
      * The ID that should be ignored.
@@ -35,85 +21,6 @@ class Unique
     protected $idColumn = 'id';
 
     /**
-     * There extra where clauses for the query.
-     *
-     * @var array
-     */
-    protected $wheres = [];
-
-    /**
-     * The custom query callback.
-     *
-     * @var \Closure|null
-     */
-    protected $using;
-
-    /**
-     * Create a new unique rule instance.
-     *
-     * @param  string  $table
-     * @param  string  $column
-     * @return void
-     */
-    public function __construct($table, $column = 'NULL')
-    {
-        $this->table = $table;
-        $this->column = $column;
-    }
-
-    /**
-     * Set a "where" constraint on the query.
-     *
-     * @param  string  $column
-     * @param  string  $value
-     * @return $this
-     */
-    public function where($column, $value = null)
-    {
-        if ($column instanceof Closure) {
-            return $this->using($column);
-        }
-
-        $this->wheres[] = compact('column', 'value');
-
-        return $this;
-    }
-
-    /**
-     * Set a "where not" constraint on the query.
-     *
-     * @param  string  $column
-     * @param  string  $value
-     * @return $this
-     */
-    public function whereNot($column, $value)
-    {
-        return $this->where($column, '!'.$value);
-    }
-
-    /**
-     * Set a "where null" constraint on the query.
-     *
-     * @param  string  $column
-     * @return $this
-     */
-    public function whereNull($column)
-    {
-        return $this->where($column, 'NULL');
-    }
-
-    /**
-     * Set a "where not null" constraint on the query.
-     *
-     * @param  string  $column
-     * @return $this
-     */
-    public function whereNotNull($column)
-    {
-        return $this->where($column, 'NOT_NULL');
-    }
-
-    /**
      * Ignore the given ID during the unique check.
      *
      * @param  mixed  $id
@@ -126,41 +33,6 @@ class Unique
         $this->idColumn = $idColumn;
 
         return $this;
-    }
-
-    /**
-     * Register a custom query callback.
-     *
-     * @param  \Closure $callback
-     * @return $this
-     */
-    public function using(Closure $callback)
-    {
-        $this->using = $callback;
-
-        return $this;
-    }
-
-    /**
-     * Format the where clauses.
-     *
-     * @return string
-     */
-    protected function formatWheres()
-    {
-        return collect($this->wheres)->map(function ($where) {
-            return $where['column'].','.$where['value'];
-        })->implode(',');
-    }
-
-    /**
-     * Get the custom query callbacks for the rule.
-     *
-     * @return array
-     */
-    public function queryCallbacks()
-    {
-        return $this->using ? [$this->using] : [];
     }
 
     /**
